Add Ranking (Legacy) protocol
This protocol is the original version of the Ranking protocol, which was
used before NEX 3.x

The protocol is largely undocumented, due to the servers having disabled
the Debug protocol and the lack of code symbols for the 3DS (which are
where the protocol is used). Because of that, some parameter types may
be inaccurate.

The only references we have are the Mario Kart 7 symbols, and we have
various method names thanks to them.