fix(messaging/types): Handle types versioning
The messaging types were assuming a version belonging to NEX 4.0 or
later. However, the types have various differences in older versions,
like the recipient ID not being split or the UserMessage inlining the
MessageRecipient fields.

In order to solve these differences, store all types in the structures
and let the implementations handle the version differences. Also
register the UserMessage derived types as object holders inheriting
Data.