feat: Implement legacy Ranking
Implement Ranking protocol used on NEX versions 1 and 2. Since the
official name of this protocol is still `Ranking`, place it inside the
folder of the NEX 3 Ranking protocol under the `legacy` name.

NEX 2 implements the category parameters as a UInt32, but NEX 1 does it
as a list of UInt16s. However, NEX 1 onlt accepts a single value inside
the list, so implement it in the API as a UInt32 and wrap around the NEX
1 cases to extract the first value from the list.

The method IDs on NEX 1 and NEX 2 don't exactly match. To account for
this difference, check for the NEX version and use different switch
cases of method IDs depending on it. The methods that have a different
ID on NEX 1 also have a constant with this ID with the suffix `NEX1`.